+++ src/net/ncplanner/plannerator/multiblock/editor/Decal.java
@@ -1,5 +1,5 @@
 package net.ncplanner.plannerator.multiblock.editor;
-import net.ncplanner.plannerator.Renderer;
+import net.ncplanner.plannerator.graphics.Renderer;
 public abstract class Decal{
     public final int x;
     public final int y;
@@ -9,7 +9,7 @@ public abstract class Decal{
         this.y = y;
         this.z = z;
     }
-    public abstract void render(Renderer renderer, double x, double y, double blockSize);
-    public abstract void render3D(Renderer renderer, double x, double y, double z, double blockSize);
+    public abstract void render(Renderer renderer, float x, float y, float blockSize);
+    public abstract void render3D(Renderer renderer, float x, float y, float z, float blockSize);
     public abstract String getTooltip();
 }
